// web/src/app/admin/passists/interfaces.ts
import { DocumentSet } from "@/lib/types";

export interface Prompt {
  id: number;
  name: string;
  shared: boolean;
  description: string;
  system_prompt: string;
  task_prompt: string;
  include_citations: boolean;
  datetime_aware: boolean;
  default_prompt: boolean;
}

export interface Passist {
  id: number;
  name: string;
  shared: boolean;
  is_visible: boolean;
  display_priority: number | null;
  description: string;
  document_sets: DocumentSet[];
  prompts: Prompt[];
  num_chunks?: number;
  llm_relevance_filter?: boolean;
  llm_filter_extraction?: boolean;
  llm_model_version_override?: string;
  default_passist: boolean;
}